requirements.txt
streamlit
rdkit
pandas
pillow
python-dotenv
google-generativeai
requests
